feat(navigation): implement stack and tab navigators
- Create stack and tab navigators to manage screen transitions.
- Add resolveScreenComponent utility for screen resolution.
- Update context exports to include new navigators.
- Refactor navigation context to support new navigator structures.
